# services/file-service/src/main/resources/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SHNEA 파일 서비스 API",
    "version": "0.1.0",
    "description": "서버 API 키와 관리자 JWT 경로를 분리한다. 브라우저에는 서버 키를 넣지 않는다. 관리자 업로드는 생성한 관리자 소유. 이용자 JWT·공유·미리보기·보존 자동 삭제는 후속 단계다."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/files/admin/environments/{environmentId}": {
      "get": {
        "operationId": "admin_get_files",
        "summary": "관리자: 같은 프로젝트·환경의 완료 파일 목록",
        "tags": [
          "Files"
        ],
        "security": [
          {
            "Administrator": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileInfo"
                  }
                }
              }
            }
          },
          "default": {
            "description": "한국어 오류. code/requestId로 처리합니다.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "environmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ]
      }
    },
    "/api/v1/files/admin/environments/{environmentId}/uploads": {
      "post": {
        "operationId": "admin_post_files_uploads",
        "summary": "관리자: 업로드 세션 생성 — requestId 재시도는 같은 세션",
        "tags": [
          "Files"
        ],
        "security": [
          {
            "Administrator": []
          }
        ],
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Upload"
                }
              }
            }
          },
          "default": {
            "description": "한국어 오류. code/requestId로 처리합니다.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCreate"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "environmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "get": {
        "operationId": "admin_resumable_uploads",
        "summary": "현재 관리자가 만든 미완료 업로드 목록",
        "security": [
          {
            "Administrator": []
          }
        ],
        "parameters": [
          {
            "name": "environmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resumable"
                  }
                }
              }
            }
          },
          "default": {
            "description": "한국어 오류",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/files/admin/environments/{environmentId}/uploads/{id}": {
      "get": {
        "operationId": "admin_get_files_uploads_id",
        "summary": "관리자: 수신 위치·만료·완료 상태 조회",
        "tags": [
          "Files"
        ],
        "security": [
          {
            "Administrator": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Upload"
                }
              }
            }
          },
          "default": {
            "description": "한국어 오류. code/requestId로 처리합니다.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "environmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "patch": {
        "operationId": "admin_patch_files_uploads_id",
        "summary": "관리자: 현재 수신 위치부터 최대 8MiB 조각 전송",
        "tags": [
          "Files"
        ],
        "security": [
          {
            "Administrator": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Upload"
                }
              }
            }
          },
          "default": {
            "description": "한국어 오류. code/requestId로 처리합니다.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "environmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Upload-Offset",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "X-Chunk-SHA256",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{64}$"
            }
          },
          {
            "name": "Content-Length",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 8388608
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "admin_delete_files_uploads_id",
        "summary": "관리자: 미완료 업로드 취소 — 완료 파일은 보존",
        "tags": [
          "Files"
        ],
        "security": [
          {
            "Administrator": []
          }
        ],
        "responses": {
          "204": {
            "description": "성공"
          },
          "default": {
            "description": "한국어 오류. code/requestId로 처리합니다.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "environmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/files/admin/environments/{environmentId}/uploads/{id}/complete": {
      "post": {
        "operationId": "admin_post_files_uploads_id_complete",
        "summary": "관리자: 전체 SHA-256 확인 후 파일 확정 — 재시도 안전",
        "tags": [
          "Files"
        ],
        "security": [
          {
            "Administrator": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileInfo"
                }
              }
            }
          },
          "default": {
            "description": "한국어 오류. code/requestId로 처리합니다.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "environmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/files/admin/environments/{environmentId}/{id}": {
      "delete": {
        "operationId": "admin_delete_files_id",
        "summary": "관리자: 파일 삭제 — 즉시 접근 차단, 물리 정리는 재시도",
        "tags": [
          "Files"
        ],
        "security": [
          {
            "Administrator": []
          }
        ],
        "responses": {
          "204": {
            "description": "성공"
          },
          "default": {
            "description": "한국어 오류. code/requestId로 처리합니다.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "environmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/files/admin/environments/{environmentId}/{id}/visibility": {
      "put": {
        "operationId": "admin_put_files_id_visibility",
        "summary": "관리자: 파일 공개·비공개 변경",
        "tags": [
          "Files"
        ],
        "security": [
          {
            "Administrator": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileInfo"
                }
              }
            }
          },
          "default": {
            "description": "한국어 오류. code/requestId로 처리합니다.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "environmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Visibility"
              }
            }
          }
        }
      }
    },
    "/api/v1/files/admin/environments/{environmentId}/{id}/download-ticket": {
      "post": {
        "operationId": "admin_download_ticket",
        "summary": "최대 60초·1회용 원본 다운로드 발급",
        "security": [
          {
            "Administrator": []
          }
        ],
        "parameters": [
          {
            "name": "environmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadTicket"
                }
              }
            }
          },
          "default": {
            "description": "한국어 오류",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ServerApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Platform-Key"
      },
      "Administrator": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "관리자 realm 발급, audience platform-admin-api, platform-admin 역할 필요"
      }
    },
    "schemas": {
      "UploadCreate": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "originalName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "size": {
            "type": "integer",
            "minimum": 0,
            "maximum": 5000000000
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ],
            "default": "PUBLIC"
          },
          "retentionCode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60,
            "default": "default"
          }
        },
        "required": [
          "requestId",
          "originalName",
          "size",
          "sha256"
        ]
      },
      "Upload": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "UPLOADING",
              "READY",
              "CANCELLED",
              "EXPIRED",
              "DELETED"
            ]
          },
          "size": {
            "type": "integer",
            "minimum": 0,
            "maximum": 5000000000
          },
          "receivedBytes": {
            "type": "integer",
            "minimum": 0
          },
          "maxChunkBytes": {
            "type": "integer",
            "const": 8388608
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "fileId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "uploadId",
          "state",
          "size",
          "receivedBytes",
          "maxChunkBytes",
          "expiresAt",
          "fileId"
        ]
      },
      "FileInfo": {
        "type": "object",
        "properties": {
          "fileId": {
            "type": "string",
            "format": "uuid"
          },
          "originalName": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "minimum": 0,
            "maximum": 5000000000
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "retentionCode": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time"
          },
          "downloadUrl": {
            "type": "string"
          }
        },
        "required": [
          "fileId",
          "originalName",
          "size",
          "sha256",
          "visibility",
          "retentionCode",
          "createdAt",
          "lastUsedAt",
          "downloadUrl"
        ]
      },
      "Visibility": {
        "type": "object",
        "properties": {
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          }
        },
        "required": [
          "visibility"
        ]
      },
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "requestId"
        ]
      },
      "Resumable": {
        "type": "object",
        "properties": {
          "upload": {
            "$ref": "#/components/schemas/Upload"
          },
          "originalName": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "retentionCode": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "upload",
          "originalName",
          "sha256",
          "visibility",
          "retentionCode",
          "requestId"
        ]
      },
      "DownloadTicket": {
        "type": "object",
        "properties": {
          "downloadUrl": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "downloadUrl",
          "expiresAt"
        ]
      }
    }
  }
}